fixed error

+++ malloc.h
@@ -1,5 +1,5 @@
 #ifndef ADD_H_GUARD
 #define ADD_H_GUARD
-void *malloc(unsigned int size);
+void *malloc(size_t size);
 void free(void* addr);
 #endif
